requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
-e git+https://github.com/edx/django-pyfs.git@1.0.3#egg=django-pyfs==1.0.3  # via -r requirements/test.in
appdirs==1.4.4            # via -r requirements/base.txt, fs
astroid==2.4.2            # via -r requirements/test.in, edx-lint, pylint, pylint-celery
attrs==20.3.0             # via pytest
bleach==3.2.1             # via -r requirements/base.txt, lti-consumer-xblock
certifi==2020.12.5        # via -r requirements/base.txt, requests
chardet==4.0.0            # via -r requirements/base.txt, requests
click-log==0.1.8          # via edx-lint
click==7.1.2              # via click-log, edx-lint
coverage==5.3.1           # via coveralls
coveralls==3.0.0          # via -r requirements/test.in
ddt==1.4.1                # via -r requirements/test.in
django-filter==2.4.0      # via -r requirements/base.txt, lti-consumer-xblock
django==2.2.17            # via -c requirements/constraints.txt, -r requirements/base.txt, django-filter, edx-opaque-keys, lti-consumer-xblock, xblock-sdk
docopt==0.6.2             # via coveralls
edx-lint==0.5.3           # via -r requirements/test.in
edx-opaque-keys[django]==2.1.1  # via -r requirements/base.txt, lti-consumer-xblock
fs==2.4.11                # via -r requirements/base.txt, xblock
future==0.18.2            # via -r requirements/base.txt, pyjwkest
idna==2.10                # via -r requirements/base.txt, requests
iniconfig==1.1.1          # via pytest
isort==5.7.0              # via pylint
lazy-object-proxy==1.4.3  # via astroid
lazy==1.4                 # via -r requirements/base.txt, lti-consumer-xblock
lti-consumer-xblock==2.4.0  # via -r requirements/base.txt
lxml==4.6.2               # via -r requirements/base.txt, lti-consumer-xblock, xblock
mako==1.1.3               # via -r requirements/base.txt, lti-consumer-xblock, xblock-utils
markupsafe==1.1.1         # via -r requirements/base.txt, mako, xblock
mccabe==0.6.1             # via pylint
mock==3.0.5               # via -c requirements/constraints.txt, -r requirements/test.in
oauthlib==3.1.0           # via -r requirements/base.txt, lti-consumer-xblock
packaging==20.8           # via -r requirements/base.txt, bleach, pytest
pbr==5.5.1                # via -r requirements/base.txt, stevedore
pep8==1.7.1               # via -r requirements/test.in
pluggy==0.13.1            # via pytest
py==1.10.0                # via pytest
pycryptodomex==3.9.9      # via -r requirements/base.txt, lti-consumer-xblock, pyjwkest
pyjwkest==1.4.2           # via -r requirements/base.txt, lti-consumer-xblock
pylint-celery==0.3        # via edx-lint
pylint-django==0.11.1     # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.6.0             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pymongo==3.11.2           # via -r requirements/base.txt, edx-opaque-keys
pyparsing==2.4.7          # via -r requirements/base.txt, packaging
pytest-django==4.1.0      # via -r requirements/test.in
pytest==6.2.1             # via pytest-django
python-dateutil==2.8.1    # via -r requirements/base.txt, xblock
pytz==2020.5              # via -r requirements/base.txt, django, fs, xblock
pyyaml==5.3.1             # via -r requirements/base.txt, xblock
requests==2.25.1          # via -r requirements/base.txt, coveralls, pyjwkest
simplejson==3.17.2        # via -r requirements/base.txt, xblock-utils
six==1.15.0               # via -r requirements/base.txt, astroid, bleach, edx-lint, edx-opaque-keys, fs, mock, pyjwkest, python-dateutil, xblock
sqlparse==0.4.1           # via -r requirements/base.txt, django
stevedore==3.3.0          # via -r requirements/base.txt, edx-opaque-keys
toml==0.10.2              # via pylint, pytest
urllib3==1.26.2           # via -r requirements/base.txt, requests
web-fragments==0.3.2      # via -r requirements/base.txt, xblock, xblock-utils
webencodings==0.5.1       # via -r requirements/base.txt, bleach
webob==1.8.6              # via -r requirements/base.txt, xblock
wrapt==1.12.1             # via astroid
xblock-sdk==0.2.2         # via -r requirements/test.in
xblock-utils==2.1.2       # via -r requirements/base.txt, lti-consumer-xblock
xblock==1.4.0             # via -r requirements/base.txt, lti-consumer-xblock, xblock-utils

# The following packages are considered to be unsafe in a requirements file:
# setuptools
